move definitions and base class from concrete cover check to eurocode/nominal_cover_constants.py
