--- blueprints/checks/nominal_concrete_cover/constants/__init__.py
@@ -1,1 +1,0 @@
-"""Collection of constants for the calculation of nominal concrete cover."""
