+++ blueprints/checks/nominal_concrete_cover/nominal_concrete_cover.py
@@ -3,13 +3,6 @@ r"""Calculation of nominal concrete cover from NEN-EN 1992-1-1: Chapter 4 - Dura
 from dataclasses import dataclass, field
 from typing import Literal
 
-from blueprints.checks.nominal_concrete_cover.constants.base import (
-    AbrasionClass,
-    CastingSurface,
-)
-from blueprints.checks.nominal_concrete_cover.constants.base import (
-    NominalConcreteCoverConstantsBase as ConstantsBase,
-)
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.formula_4_1 import Form4Dot1NominalConcreteCover
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.formula_4_2 import Form4Dot2MinimumConcreteCover
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.table_4_1 import (
@@ -23,6 +16,13 @@ from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.t
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.table_4_2 import Table4Dot2MinimumCoverWithRegardToBond
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.table_4_4n import (
     Table4Dot4nMinimumCoverDurabilityReinforcementSteel,
+)
+from blueprints.codes.eurocode.nominal_cover_constants import (
+    AbrasionClass,
+    CastingSurface,
+)
+from blueprints.codes.eurocode.nominal_cover_constants import (
+    NominalConcreteCoverConstantsBase as ConstantsBase,
 )
 from blueprints.codes.eurocode.structural_class import ConcreteStructuralClassBase
 from blueprints.codes.latex_formula import latex_max_curly_brackets
