+++ tests/checks/nominal_concrete_cover/test_nominal_concrete_cover.py
@@ -2,7 +2,6 @@
 
 import pytest
 
-from blueprints.checks.nominal_concrete_cover.constants.base import AbrasionClass, CastingSurface, NominalConcreteCoverConstantsBase
 from blueprints.checks.nominal_concrete_cover.nominal_concrete_cover import NominalConcreteCover
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.constants import NominalConcreteCoverConstants
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.table_4_1 import (
@@ -14,6 +13,7 @@ from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.t
     Table4Dot1ExposureClasses,
 )
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.table_4_3 import Table4Dot3ConcreteStructuralClass
+from blueprints.codes.eurocode.nominal_cover_constants import AbrasionClass, CastingSurface, NominalConcreteCoverConstantsBase
 from blueprints.codes.eurocode.structural_class import ConcreteStructuralClassBase
 from blueprints.materials.concrete import ConcreteMaterial
 from blueprints.type_alias import MM
