+++ tests/codes/eurocode/en_1992_1_1_2004/chapter_4_durability_and_cover/test_constants.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from blueprints.checks.nominal_concrete_cover.constants.base import AbrasionClass, CastingSurface
 from blueprints.codes.eurocode.en_1992_1_1_2004.chapter_4_durability_and_cover.constants import NominalConcreteCoverConstants
+from blueprints.codes.eurocode.nominal_cover_constants import AbrasionClass, CastingSurface
 from blueprints.type_alias import MM
